+++ packages/sources/src/components/table/table.models.ts
@@ -10,4 +10,5 @@ export interface Table<TemplateFnReturnType> {
   content: TableContent<TemplateFnReturnType>;
   headingColumns?: boolean;
   role?: string;
+  verticalLines?: boolean;
 }